EventFactory is dead, long live EventFactory!
