+++ logstash-core/src/main/java/org/logstash/plugins/ContextImpl.java
@@ -23,11 +23,7 @@ package org.logstash.plugins;
 import co.elastic.logstash.api.*;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-import org.logstash.ConvertedMap;
 import org.logstash.log.DefaultDeprecationLogger;
-
-import java.io.Serializable;
-import java.util.Map;
 
 /**
  * Default implementation of Logstash Plugin's {@link Context}
@@ -68,19 +64,6 @@ public class ContextImpl implements Context {
 
     @Override
     public EventFactory getEventFactory() {
-        return new EventFactory() {
-            @Override
-            public Event newEvent() {
-                return new org.logstash.Event();
-            }
-
-            @Override
-            public Event newEvent(Map<? extends Serializable, Object> data) {
-                if (data instanceof ConvertedMap) {
-                    return new org.logstash.Event((ConvertedMap)data);
-                }
-                return new org.logstash.Event(data);
-            }
-        };
+        return BasicEventFactory.INSTANCE;
     }
 }
